a slightly less horrible implementation of error handling
Simplified Parser and Printer types to be just simple functions, with Parsers producing a single best failure and a list of potential results/remainders. This seems to work and isn't hideous, but threading the values is relatively painful compared to the nice, clean version in the paper which only have to worry about successful results.