rename/rework the types around Syntax
The idea here is to make Syntax, Parser, and Printer, all name the
things that are most visible to the user of the library: `Syntax` is
what you define, and `Parser` and `Printer` are the types of the
functions you make from it.

To make that work, the value that provides the combinators is now an
explicit parameter. That seems more natural but necessitated moving
things around to make the DSL work.

Closes #9.