rework type names a bit
`Syntax` is now an abstract class that is to be implemented for each
syntax, capturing the type of values is can parse/print. It also
provides `parse` and `print` as simple methods which can be used
directly.

`Transcriber` is the new name for the typeclass that is implemented the
Parser and Printer types.

`Parser` and `Printer` are no longer part of the API that is seen by
the Syntax implementor, so `Parser` is now called `PartialParser` to
hopefully clarify its role slightly.